MP3 Player (Ahmed Mahfouz)
I'm writing an application that will use the JLayer library to play an
mp3 file. Since the player does not have any methods to pause the
playback - play() does not return until the song is finished - I decided
to put it in a thread and simply control that thread. This is possible
because one can execute play(1) which plays a single frame of the song.
Putting this in a loop will thus play the entire song smoothly.